% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-manipulation.R
\name{archive}
\alias{archive}
\title{Slice data_frame into files}
\usage{
archive(df, tz = "UTC", path = "\%Y-\%m.dat")
}
\arguments{
\item{df}{data_frame to save. Must contain a column named 'time'}

\item{tz}{timezone of POSIXct timestamp in df}

\item{path}{file naming scheme used to split and save data. Format must be
compatible with \code{strptime}}
}
\description{
\code{archive} splits data_frame by time specified in the \code{path}
  argument and saves to files.
}

