% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uataq-tools.R
\name{bad_data}
\alias{bad_data}
\title{Generate bad data strings}
\usage{
bad_data(t_start, t_end, comment, miu_old = "all", miu_new = NA)
}
\arguments{
\item{t_start}{character timestamp to begin changing data}

\item{t_end}{character timestamp to end changing data}

\item{comment}{comment about why the data is being removed, typically 
formatted as 'Name: here is my comment'}

\item{miu_old}{character of the ID string to remove or 'all' to specify all 
valve positions}

\item{miu_new}{character string to replace the old miu values with, or NA to
remove the data}
}
\description{
\code{bad_data} takes the given information and formats a string compatible
  with UATAQ processing routines. Times should be in MST/MDT (as displayed
  on the UATAQ data viewer) and formatted as '2016-02-12 00:00:00.000'.
}

