% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.r
\name{bootstrap}
\alias{bootstrap}
\title{Bootstrap resampling}
\usage{
bootstrap(x, fun, size = nrow(x), iter = 10)
}
\arguments{
\item{x}{data frame to be resampled}

\item{fun}{function that takes a single argument (a data frame of each
resampled dataset) and returns a single value}

\item{size}{size of each sample}

\item{iter}{number of bootstrap iterations}
}
\description{
\code{bootstrap} resamples the dataset with replacement and applies a
  function to each sample
}
