% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/br.r
\name{br}
\alias{br}
\title{Send newline characters to the R console}
\usage{
br(n = 1)
}
\arguments{
\item{n}{number of newline characters to print; negative n will clear the
console with uataq::clc()}
}
\description{
Send newline characters to the R console
}

