% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-manipulation.R
\name{breakstr}
\alias{breakstr}
\title{Split delimited strings}
\usage{
breakstr(strings, pattern = ",", ncol = NULL)
}
\arguments{
\item{strings}{character array of strings to be split}

\item{pattern}{delimiter at which to split strings}

\item{ncol}{number of columns. NULL will determine by the median number
of columns found in strings}
}
\description{
\code{breakstr} splits strings by the given delimiter and returns a
  data_frame. Lines that do not fit the correct number of columns
  are omitted.
}

