% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.r
\name{calibrate}
\alias{calibrate}
\title{Calibrate trace gas data}
\usage{
calibrate(time, gasm, gask, auto = F, er_tol = 0.1, dt_tol = 18000)
}
\arguments{
\item{time}{POSIXct class timestamp, length n}

\item{gasm}{numeric measured values, length n}

\item{gask}{numeric gas concentration flag, length n. -10 for atmospheric
observations, -99 for flush periods, or a positive value if
sampling a known concentration}

\item{auto}{if TRUE, stops calibration if the dataset ends with reference gas
flow still active}

\item{er_tol}{tolerated deviation from known value. e.g. 0.1 for 10 percent
deviation from known}

\item{dt_tol}{tolerated amount of time between observations, in seconds. If
exceeded, will remove all atmospheric data until the next
calibration}
}
\description{
\code{calibrate} atmospheric trace gas data against several reference gas
  tanks.
}
