% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorize.r
\name{colorize}
\alias{colorize}
\title{Generates hex colors for plotting}
\usage{
colorize(data, clim = NULL, ncol = 64, cols = c("blue", "cyan",
  "green", "yellow", "orange", "red"), ...)
}
\arguments{
\item{data}{data to be converted to color scale}

\item{clim}{values in data to limit color scale, as c(min, max)}

\item{ncol}{number of colors to use between cmin and cmax}

\item{cols}{colors to interpolate between}

\item{...}{arguments to pass to colorRampPalette}
}
\description{
\code{colorize} uses colorRampPalette() to generate custom color scales for
plotting.
}
