% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_lines.r
\name{count_lines}
\alias{count_lines}
\title{Count lines in a file}
\usage{
count_lines(file)
}
\arguments{
\item{file}{path to file in question}
}
\description{
\code{count_lines} uses UNIX system command \code{wc} to determine the number
  of lines in the given file
}
