% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_neighbor.r
\name{find_neighbor}
\alias{find_neighbor}
\title{Find nearest neighbors}
\usage{
find_neighbor(x, y)
}
\arguments{
\item{x}{desired values}

\item{y}{table to search values}
}
\description{
\code{find_neighbor} finds the closest value index in y for each x,
  using findInterval. Returns the index to be applied to y that most
  closely matches x.
}
