% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/findNeighbor.R
\name{findNeighbor}
\alias{findNeighbor}
\title{Find nearest neighbors}
\usage{
findNeighbor(x, y)
}
\arguments{
\item{x}{desired values}

\item{y}{table to search values}
}
\description{
\code{findNeighbor} finds the closest value index in y for each x, using findInterval
}

