% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps-tools.R
\name{gps_dir}
\alias{gps_dir}
\title{Calculate GPS direction}
\usage{
gps_dir(lat, lon)
}
\arguments{
\item{lat}{latitude, in dd.dddd}

\item{lon}{longitude, in dd.dddd}
}
\description{
\code{gps_dir} calculates the direction between GPS coordinates
}

