% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps-tools.R
\name{gps_distance}
\alias{gps_distance}
\title{Calculate GPS distances}
\usage{
gps_distance(lat, lon, units = "m")
}
\arguments{
\item{lat}{latitude, in dd.dddd}

\item{lon}{longitude, in dd.dddd}

\item{units}{output units, either 'm' or 'rad'}
}
\description{
\code{gps_distance} calculates the distances between GPS coordinates
}

