% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps_dm2dd.r
\name{gps_dm2dd}
\alias{gps_dm2dd}
\title{Convert dddmm.mmm to dd.dddd coordinates}
\usage{
gps_dm2dd(x)
}
\arguments{
\item{x}{numeric coordinate, either latitude or longitude, to convert}
}
\description{
\code{gps_dm2dd} converts the raw GPS $GPGGA output in the form of
  dddmm.mmm to map compatible degree form dd.dddd.
}

