% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps_move.r
\name{gps_move}
\alias{gps_move}
\title{Calculate GPS speed, distance, and direction}
\usage{
gps_move(Time, lat, lon)
}
\arguments{
\item{Time}{POSIXct times for observations}

\item{lat}{latitude, in dd.dddd}

\item{lon}{longitude, in dd.dddd}
}
\value{
list containing (1) \code{$speed} at each point, (2)
  \code{$distance} from the previous point, and (3) \code{$direction}
  the direction of movement in degrees.
}
\description{
\code{gps_move} calculates the speeds, distances, and directions given
  vectors of GPS coordinates
}

