% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps-tools.R
\name{gps_speed}
\alias{gps_speed}
\title{Calculate GPS speed}
\usage{
gps_speed(Time, lat, lon)
}
\arguments{
\item{Time}{POSIXct times for observations}

\item{lat}{latitude, in dd.dddd}

\item{lon}{longitude, in dd.dddd}
}
\description{
\code{gps_speed} calculates the speed in m/s between GPS coordinates
}

