% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization-tools.R
\name{iggplot}
\alias{iggplot}
\title{Generates interactive ggplot}
\usage{
iggplot(fig)
}
\arguments{
\item{fig}{ggplot object to be made interactive}
}
\description{
\code{iggplot} starts new interactive gadget session to create and view
  the given ggplot object. Returns the zoomed ggplot object when the 
  'Done' button is pressed.
}
\examples{
x <- seq(0, 2*pi, length.out=100)
f <- qplot(x, sin(x))
iggplot(f)
                 
}

