% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keeling.r
\name{keeling}
\alias{keeling}
\title{Keeling isotope analysis}
\usage{
keeling(co2, d13c, bg_co2 = 400, bg_d13c = -8)
}
\arguments{
\item{co2}{numeric co2 concentrations, in ppm}

\item{d13c}{numeric d13C, in per mille}

\item{bg_co2}{numeric background co2 concentration, in ppm}

\item{bg_d13c}{numeric background d13C, in per mille}
}
\value{
list containing (1) \code{$fit} the linear fit object and (2)
  \code{$fig} the ggplot figure.
}
\description{
\code{keeling} performs keeling isotope analysis, generating a linear fit
  and d13C keeling plot.
}

