% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keeling_split.r
\name{keeling_split}
\alias{keeling_split}
\title{Keeling isotope analysis}
\usage{
keeling_split(time, co2, d13c, format = "\%Y-\%m-01", tz = "UTC")
}
\arguments{
\item{time}{POSIXct timestamps of measurements}

\item{co2}{numeric co2 concentrations, in ppm}

\item{d13c}{numeric d13C, in per mille}

\item{format}{character format compatible with strptime to split the data
by. Defaults to '\%Y-\%m-01' which indicates monthly regression statistics
with floored timestamps (e.g. 2016-01-01 indicates the statistics for the
month of January, 2016)}
}
\value{
data_frame containing the slope, yintercept, fit statistics including
  \code{r_squared}, \code{rmse}, \code{std_error}, and \code{n}, and
  \code{lm_obj} the fit object created using the lm function for each time
  split
}
\description{
\code{keeling} performs keeling isotope analysis, generating a linear fit
  using QR regression and a d13C keeling plot.
}

