% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_interp.r
\name{na_interp}
\alias{na_interp}
\title{Linearly interpolate NA values}
\usage{
na_interp(y, x = NULL)
}
\arguments{
\item{y}{numeric vector in which to fill bracketed NA values}

\item{x}{vector giving dimension of y. NULL if y is equally spaced}
}
\description{
\code{na_interp} linearly interpolates NA values found in vector y with
  respect to dimension x (e.g. timestamp).
}

