% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rds2csv.R
\name{rds2csv}
\alias{rds2csv}
\title{Running mean smoothing}
\usage{
rds2csv(rds, file, sep = ",", quote = F, ...)
}
\arguments{
\item{rds}{path to .rds file.}

\item{file}{either a character string naming a file or a connection open for writing. '' indicates output to the console.}

\item{sep}{the field separator string.}

\item{quote}{a logical value (TRUE or FALSE) or a numeric vector. If TRUE, any character or factor columns will be surrounded by double quotes. If a numeric vector, its elements are taken as the indices of columns to quote. In both cases, row and column names are quoted if they are written. If FALSE, nothing is quoted.}

\item{...}{arguments passed to write.table}
}
\description{
\code{runSmooth} generates smoothed representation of x
}

