% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_smooth.r
\name{run_smooth}
\alias{run_smooth}
\title{Running mean smoothing}
\usage{
run_smooth(x, n = 10)
}
\arguments{
\item{x}{numeric values}

\item{n}{number of points to smooth by}
}
\description{
\code{run_smooth} generates smoothed representation of x
}
