% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-manipulation.R
\name{runSmooth}
\alias{runSmooth}
\title{Running mean smoothing}
\usage{
runSmooth(x, n = 10)
}
\arguments{
\item{x}{numeric values}

\item{n}{number of points to smooth by}
}
\description{
\code{runSmooth} generates smoothed representation of x
}

