% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stilt_init.r
\name{stilt_init}
\alias{stilt_init}
\title{Initialize new STILT project}
\usage{
stilt_init(project, branch = "master", repo = "https://github.com/uataq/stilt")
}
\arguments{
\item{project}{name or location to initialize new project}

\item{branch}{name of repo branch to checkout}

\item{repo}{repo to fetch stilt from using \code{git clone repo}}
}
\description{
Initialize new STILT project
}
\author{
Ben Fasoli

\code{stilt_init} generates the framework for a new STILT modeling project
}
