% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.r
\name{archive}
\alias{archive}
\title{Slice data_frame into files}
\usage{
archive(df, tz = "UTC", path = "\%Y_\%m.dat", col_names = NULL)
}
\arguments{
\item{df}{data_frame to save. Must contain a column named 'time'}

\item{tz}{timezone of POSIXct timestamp in df}

\item{path}{file naming scheme used to split and save data. Format must be
compatible with \code{strptime}}

\item{col_names}{whether to write header line with column names, NULL applies
column names to new files only, FALSE doesn't write a header, and TRUE
will write a line with the column names even if appending. Be careful with
TRUE, as you may end up with column flags mid-file.}
}
\description{
\code{archive} splits data_frame by time specified in the \code{path}
  argument and saves to files.
}
