% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stilt_init.r
\name{stilt_init}
\alias{stilt_init}
\title{Initialize new STILT project}
\usage{
stilt_init(name, repo = "https://github.com/benfasoli/stilt")
}
\arguments{
\item{name}{name or location to initialize new project}

\item{repo}{}
}
\description{
Initialize new STILT project
}
\author{
Ben Fasoli

\code{stilt_init} generates the framework for a new STILT modeling project
}

